requirements.txt
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
icalendar==5.0.13
idna==3.8
python-dateutil==2.9.0.post0
python-telegram-bot==21.5
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
urllib3==2.2.3